Move titan text to embedding into TextToEmbedding
Now we only have one embedding implementation we no longer need the titan namespacing.

This removes the titan namespacing from the text to embedding code and
moves the logic directly into the TextToEmbedding class.
