+++ spec/lib/tasks/search_spec.rb
@@ -71,7 +71,7 @@ RSpec.describe "rake search tasks" do
     before do
       Rake::Task[task_name].reenable
 
-      allow(Search::TextToEmbedding::Titan).to receive(:call) do |arg|
+      allow(Search::TextToEmbedding).to receive(:call) do |arg|
         arg.map { |a| mock_titan_embedding(a) }
       end
     end
